feat(classroom): up/down reorder for teacher items in the curriculum tab
Teachers can now reorder their own classroom assignments inside a
curriculum unit using up/down arrows. Admin lessons / projects /
curriculum assignments stay in their globally-fixed order.

Schema (012_curriculum_order_numeric.sql):
- assignments.curriculum_order changed from integer to numeric so
  teacher items can slot between admin integers via half-step values

Backend:
- AssignmentCreate / AssignmentUpdate types now accept float

Frontend:
- Teacher curriculum tab: green up/down arrows on every 'YOURS' row,
  disabled at the top/bottom of the unit
- Move logic computes a midpoint between the adjacent items'
  effective order values, so teacher assignments cleanly interleave
  with admin content without ever touching admin's order_index

+++ backend/routers/assignments.py
@@ -34,7 +34,7 @@ class AssignmentCreate(BaseModel):
     hint_1: Optional[str] = None
     hint_2: Optional[str] = None
     curriculum_unit_id: Optional[str] = None  # if set, this assignment shows in the curriculum tab of the teacher's classroom
-    curriculum_order: Optional[int] = None
+    curriculum_order: Optional[float] = None  # numeric so we can slot between admin items via half-steps
 
 
 class AssignmentUpdate(BaseModel):
@@ -49,7 +49,7 @@ class AssignmentUpdate(BaseModel):
     hint_1: Optional[str] = None
     hint_2: Optional[str] = None
     curriculum_unit_id: Optional[str] = None
-    curriculum_order: Optional[int] = None
+    curriculum_order: Optional[float] = None
 
 
 # ============================================================
